+++ Modules/Sync.lua
@@ -8,19 +8,18 @@
       sync           T - Actual sync data.
       syncAck        T - Sync ack received.
 ]]
----@type RCLootCouncil
+
 --- @type RCLootCouncil
 local addon = select(2, ...)
----@class Sync
+---@class Sync : AceSerializer-3.0
 local sync = addon:NewModule("Sync", "AceSerializer-3.0")
 local LibDialog = LibStub("LibDialog-1.0")
 local LD = LibStub("LibDeflate")
+--- @type RCLootCouncilLocale
 local L = LibStub("AceLocale-3.0"):GetLocale("RCLootCouncil")
 local AG = LibStub("AceGUI-3.0")
 local PREFIX = addon.PREFIXES.SYNC
----@type Services.Comms
 local Comms = addon.Require "Services.Comms"
----@type Data.Player
 local Player = addon.Require "Data.Player"
 addon.Sync = sync